indexer/requirements.txt
azure-search-documents==11.6.0b8
azure-storage-blob
azure-identity
python-dotenv